Improve scrolling performance a little.
Mark the table view's scroll view as layer backed, and moves some table
cell view initialization so it only runs when a cell view is created
rather than every time a cell view is requested.

This change moves the bottleneck from drawing to layout. The result is
that most scrolling takes place at over 50 fps in my testing. With
extremely aggressive scrolling it's possible to get the framerate to
drop when the amount of layout work exceeds the time the main thread has
available between frames.