Fix to wait until the first notification occurs due to sync loading the schema.
The connect: method on RLMRealmNode now offers a schemaLoadedCallback.

I moved the code that was occuring directly after connect: into the callback which means the UI doesn't load until sync loads the schema.

The problem though is that this could take awhile, so the UI should show a spinner and I assume we should have a timeout or something to inform user that the sync server is unavailable or unresponsive.